rattler_networking

Module retry_policies

Reexports the trait RetryPolicy from the retry_policies crate as well as all implementations.

This module also provides the DoNotRetryPolicy which is useful if you do not want to retry anything.

Structs

DoNotRetryPolicy

A simple RetryPolicy that just never retries.

ExponentialBackoff

Exponential backoff with optional jitter.

ExponentialBackoffBuilder

Builds an exponential backoff policy.

ExponentialBackoffTimed

Exponential backoff with a maximum retry duration.

Enums

Jitter

How to apply jitter to the retry intervals.

RetryDecision

Outcome of evaluating a retry policy for a failed task.

Traits

RetryPolicy

A policy for deciding whether and when to retry.

Functions

default_retry_policy

Returns the default retry policy that can be used.