rattler_networking::retry_policies

Struct ExponentialBackoff

#[non_exhaustive]
pub struct ExponentialBackoff {
    pub max_n_retries: Option<u32>,
    pub min_retry_interval: Duration,
    pub max_retry_interval: Duration,
    pub jitter: Jitter,
    pub base: u32,
}

Exponential backoff with optional jitter.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

max_n_retries: Option<u32>

Maximum number of allowed retries attempts.

min_retry_interval: Duration

Minimum waiting time between two retry attempts (it can end up being lower due to jitter).

max_retry_interval: Duration

Maximum waiting time between two retry attempts.

jitter: Jitter

How we apply jitter to the calculated backoff intervals.

base: u32

Base of the exponential

Implementations

impl ExponentialBackoff

pub fn builder() -> ExponentialBackoffBuilder

Returns a builder.

Example

use retry_policies::policies::ExponentialBackoff;
use std::time::Duration;

let backoff = ExponentialBackoff::builder()
    .build_with_max_retries(5);

assert_eq!(backoff.min_retry_interval, Duration::from_secs(1));
assert_eq!(backoff.max_retry_interval, Duration::from_secs(30 * 60));
assert_eq!(backoff.max_n_retries, Some(5));
assert_eq!(backoff.base, 2);