rattler_networking::retry_policies

Trait RetryPolicy

pub trait RetryPolicy {
    // Required method
    fn should_retry(
        &self,
        request_start_time: SystemTime,
        n_past_retries: u32,
    ) -> RetryDecision;
}

A policy for deciding whether and when to retry.

Required Methods

fn should_retry( &self, request_start_time: SystemTime, n_past_retries: u32, ) -> RetryDecision

Determine if a task should be retried according to a retry policy.