Crate rattler_networking

Networking utilities for Rattler, specifically authenticating requests

Re-exports

pub use authentication_middleware::AuthenticationMiddleware;

pub use authentication_storage::authentication::Authentication;

pub use authentication_storage::storage::AuthenticationStorage;

pub use mirror_middleware::MirrorMiddleware;

pub use oci_middleware::OciMiddleware;

pub use gcs_middleware::GCSMiddleware;

pub use s3_middleware::S3Middleware;

Modules

authentication_middleware

reqwest middleware that authenticates requests with data from the AuthenticationStorage

authentication_storage

This module contains the authentication storage backend trait and implementations

gcs_middleware

Middleware to handle gcs:// URLs to pull artifacts from an GCS

mirror_middleware

Middleware to handle mirrors

oci_middleware

Middleware to handle oci:// URLs to pull artifacts from an OCI registry

retry_policies

Reexports the trait RetryPolicy from the retry_policies crate as well as all implementations.

s3_middleware

Middleware to handle s3:// URLs to pull artifacts from an S3 bucket

Structs

LazyClient

Initialization of the reqwest client can be expensive, using this struct allows creating the client only when needed.